+++ src/Sulu/Bundle/MediaBundle/Resources/public/js/components/media-selection/overlay/main.js
@@ -491,7 +491,7 @@ define([
                             selectOnAction: !this.options.singleSelect,
                             unselectOnBackgroundClick: false,
                             locale: this.options.locale,
-                            hintIcon: 'fa-check-circle-o',
+                            actionIcons: ['fa-check-circle-o'],
                             badges: [
                                 {
                                     column: 'title',